wip(api-headless-cms-scheduler): action fix

+++ packages/api-scheduler/src/features/CancelScheduledAction/CancelScheduledActionUseCase.ts
@@ -5,6 +5,7 @@ import { ScheduledActionModel, SchedulerService } from "~/shared/abstractions.js
 import { ScheduledActionNotFoundError, ScheduledActionPersistenceError } from "~/domain/errors.js";
 import { DeleteEntryUseCase } from "@webiny/api-headless-cms/features/contentEntry/DeleteEntry/index.js";
 import { ScheduledActionIdWithVersion } from "~/domain/ScheduledActionIdWithVersion.js";
+import { EntryNotFoundError } from "@webiny/api-headless-cms/domain/contentEntry/errors.js";
 
 /**
  * Cancels a scheduled action
@@ -42,7 +43,13 @@ class CancelScheduledActionUseCaseImpl implements UseCaseAbstraction.Interface {
         // Delete EventBridge schedule
         // Note: We continue even if this fails, as the schedule might already be executed/deleted
         try {
-            await this.schedulerService.delete(id);
+            const eventBridgeSchedule = await this.schedulerService.exists(id);
+            /**
+             * No point to even try deleting if it doesn't exist.
+             */
+            if (eventBridgeSchedule) {
+                await this.schedulerService.delete(id);
+            }
         } catch (error) {
             console.warn(
                 `Failed to delete EventBridge schedule: ${scheduleId}. Continuing with CMS entry deletion.`,
@@ -57,6 +64,12 @@ class CancelScheduledActionUseCaseImpl implements UseCaseAbstraction.Interface {
         });
 
         if (deleteResult.isFail()) {
+            /**
+             * Some process could have already deleted the entry, in which case we can safely ignore this error.
+             */
+            if (deleteResult.error instanceof EntryNotFoundError) {
+                return Result.ok();
+            }
             return Result.fail(
                 new ScheduledActionPersistenceError(new Error(deleteResult.error.message))
             );
